Changing the organization of the code.
This commit changes how the different files are structured. At the very
top level we have the `CRRao.jl`, `fitmodel.jl` and `general_stats.jl`
files. All the definitions of the `fitmodel` function are moved to the
`src/fitmodel_defs` directory. Moreover, each file containing
definitions of `fitmodel` are responsible for including all the code
required for those definitions.